fix: spaces are not allowed to be printed anymore

+++ src/projects/8-todoDeleteV2/TodoV2/ItemForm.js
@@ -13,7 +13,7 @@ const ItemForm = (props) => {
 
   const handleSubmit = (ev) => {
     ev.preventDefault()
-    if (!item.name) return
+    if (!item.name || !item.name.replaceAll(' ', '')) return
     props.addItem(item)
     setItem(initialFormState)
   }
